+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Development version
 
+* Add support for the `License-File` field for specifying additional licensing-related files (PEP 639).
 * Ensure compatibility with Python 3.14. The changes only affect testing.
 
 # Version 0.6.0 - 2025-06-26